Refactor compilation
